Warp Drive Projector Finished UI Improvements to Items API Changes

+++ src/main/java/mod/stf/syconn/api/util/AnchorPos.java
@@ -3,7 +3,7 @@ package mod.stf.syconn.api.util;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 
-public record AnchorPos(double x, double y, double z){
+public record AnchorPos(int x, int y, int z){
 
     public BlockPos getBlockPos(){
         return new BlockPos(x, y, z);
@@ -11,13 +11,13 @@ public record AnchorPos(double x, double y, double z){
 
     public CompoundTag save(){
         CompoundTag tag = new CompoundTag();
-        tag.putDouble("x", x);
-        tag.putDouble("y", y);
-        tag.putDouble("z", z);
+        tag.putInt("x", x);
+        tag.putInt("y", y);
+        tag.putInt("z", z);
         return tag;
     }
 
     public static AnchorPos read(CompoundTag tag){
-        return new AnchorPos(tag.getDouble("x"), tag.getDouble("y"), tag.getDouble("z"));
+        return new AnchorPos(tag.getInt("x"), tag.getInt("y"), tag.getInt("z"));
     }
 }
